Refactor: made cpu logic own class/all tests pass
