+++ bin/tictactoe.rb
@@ -6,8 +6,9 @@ questions = Questions.new
 game_config = GameConfig.new(questions, parsed_arguments)
 game_logic = GameLogic.new
 console_output = ConsoleOutput.new(game_config)
+computer_logic = ComputerLogic.new(game_logic)
 
-game_loop = GameLoop.new(questions, game_logic, game_config, console_output)
+game_loop = GameLoop.new(questions, game_logic, game_config, console_output, computer_logic)
 
 game_loop.run()
